+++ addons/yat/src/commands/builtin/Echo.cs
@@ -8,14 +8,10 @@ namespace YAT.Commands
 	[Argument("message", "string", "The text to display.")]
 	public partial class Echo : ICommand
 	{
-		public YAT Yat { get; set; }
-
-		public Echo(YAT Yat) => this.Yat = Yat;
-
-		public CommandResult Execute(params string[] args)
+		public CommandResult Execute(CommandArguments args)
 		{
-			var text = string.Join(" ", args[1..^0]);
-			Yat.Terminal.Print(text);
+			var text = string.Join(" ", args.Arguments[1..^0]);
+			args.Terminal.Print(text);
 
 			return CommandResult.Success;
 		}