+++ addons/yat/src/commands/builtin/Restart.cs
@@ -7,14 +7,10 @@ namespace YAT.Commands
 	[Command("restart", "Restarts the level.", "[b]Usage[/b]: restart", "reboot")]
 	public partial class Restart : ICommand
 	{
-		public YAT Yat { get; set; }
-
-		public Restart(YAT Yat) => this.Yat = Yat;
-
-		public CommandResult Execute(params string[] args)
+		public CommandResult Execute(CommandArguments args)
 		{
-			Yat.Terminal.Print($"Restarting {Yat.GetTree().CurrentScene.Name}...");
-			Yat.GetTree().ReloadCurrentScene();
+			args.Terminal.Print($"Restarting {args.Yat.GetTree().CurrentScene.Name}...");
+			args.Yat.GetTree().ReloadCurrentScene();
 
 			return CommandResult.Success;
 		}